+++ HISTORY.md
@@ -1,6 +1,7 @@
 # Version 0.0.7
 - Moved type checks to functions and globals
 - Removed unnecessary whitespace from model formula
+- Introduced `max_width` with default to 64 characters 
 
 # Version 0.0.6
 - Introduced check if 't' column is present, otherwise use 'z' column to handle models with only one coefficient